Migrate almost all remaining tests to Kotlin

+++ stripe/src/test/java/com/stripe/android/model/AccountParamsTest.kt
@@ -9,13 +9,17 @@ class AccountParamsTest {
     @Test
     fun toParamMap_withBusinessData() {
         val businessData = mapOf("name" to "Stripe")
-        val params = AccountParams.createAccountParams(true,
-            AccountParams.BusinessType.Company, businessData)
+        val params = AccountParams
+            .createAccountParams(
+                true,
+                AccountParams.BusinessType.Company,
+                businessData
+            )
             .toParamMap()
 
         val accountData = params["account"] as Map<String, *>?
         assertNotNull(accountData)
-        assertEquals(3, accountData.size.toLong())
+        assertEquals(3, accountData.size)
         assertEquals(java.lang.Boolean.TRUE, accountData[AccountParams.API_TOS_SHOWN_AND_ACCEPTED])
         assertEquals(AccountParams.BusinessType.Company.code,
             accountData[AccountParams.API_BUSINESS_TYPE])
@@ -31,7 +35,7 @@ class AccountParamsTest {
 
         val accountData = params["account"] as Map<String, *>?
         assertNotNull(accountData)
-        assertEquals(1, accountData.size.toLong())
+        assertEquals(1, accountData.size)
         assertEquals(java.lang.Boolean.TRUE, accountData[AccountParams.API_TOS_SHOWN_AND_ACCEPTED])
     }
 }
